Moved NewAccount and SendMagicLinks to a new fragment
