+++ example/src/main/java/org/wordpress/android/fluxc/example/MainFragment.java
@@ -1,6 +1,5 @@
 package org.wordpress.android.fluxc.example;
 
-import android.app.AlertDialog;
 import android.app.DialogFragment;
 import android.app.Fragment;
 import android.app.FragmentTransaction;
@@ -11,7 +10,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.ViewGroup;
-import android.widget.EditText;
 
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
@@ -26,11 +24,9 @@ import org.wordpress.android.fluxc.network.MemorizingTrustManager;
 import org.wordpress.android.fluxc.network.discovery.SelfHostedEndpointFinder.DiscoveryError;
 import org.wordpress.android.fluxc.store.AccountStore;
 import org.wordpress.android.fluxc.store.AccountStore.AuthenticatePayload;
-import org.wordpress.android.fluxc.store.AccountStore.NewAccountPayload;
 import org.wordpress.android.fluxc.store.AccountStore.OnAccountChanged;
 import org.wordpress.android.fluxc.store.AccountStore.OnAuthenticationChanged;
 import org.wordpress.android.fluxc.store.AccountStore.OnDiscoveryResponse;
-import org.wordpress.android.fluxc.store.AccountStore.OnNewUserCreated;
 import org.wordpress.android.fluxc.store.SiteStore;
 import org.wordpress.android.fluxc.store.SiteStore.OnSiteChanged;
 import org.wordpress.android.fluxc.store.SiteStore.RefreshSitesXMLRPCPayload;
@@ -97,10 +93,13 @@ public class MainFragment extends Fragment {
             }
         });
 
-        view.findViewById(R.id.new_account).setOnClickListener(new OnClickListener() {
+        view.findViewById(R.id.signed_out_actions).setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View v) {
-                showNewAccountDialog();
+                getFragmentManager().beginTransaction()
+                        .replace(R.id.fragment_container, new SignedOutActionsFragment())
+                        .addToBackStack(null)
+                        .commit();
             }
         });
 
@@ -109,13 +108,6 @@ public class MainFragment extends Fragment {
         view.findViewById(R.id.posts).setOnClickListener(getOnClickListener(new PostsFragment()));
         view.findViewById(R.id.comments).setOnClickListener(getOnClickListener(new CommentsFragment()));
         view.findViewById(R.id.media).setOnClickListener(getOnClickListener(new MediaFragment()));
-
-        view.findViewById(R.id.magic_link_email).setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                showMagicLinkDialog();
-            }
-        });
 
         return view;
     }
@@ -251,37 +243,6 @@ public class MainFragment extends Fragment {
         mDispatcher.dispatch(SiteActionBuilder.newFetchSitesXmlRpcAction(payload));
     }
 
-    private void showNewAccountDialog() {
-        FragmentTransaction ft = getFragmentManager().beginTransaction();
-        DialogFragment newFragment = ThreeEditTextDialog.newInstance(new Listener() {
-            @Override
-            public void onClick(String username, String email, String password) {
-                newAccountAction(username, email, password);
-            }
-        }, "Username", "Email", "Password");
-        newFragment.show(ft, "dialog");
-    }
-
-    private void newAccountAction(String username, String email, String password) {
-        NewAccountPayload newAccountPayload = new NewAccountPayload(username, password, email, true);
-        mDispatcher.dispatch(AccountActionBuilder.newCreateNewAccountAction(newAccountPayload));
-    }
-
-    private void showMagicLinkDialog() {
-        AlertDialog.Builder alert = new AlertDialog.Builder(getActivity());
-        final EditText editText = new EditText(getActivity());
-        editText.setSingleLine();
-        alert.setMessage("Send magic link login e-mail:");
-        alert.setView(editText);
-        alert.setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog, int whichButton) {
-                String email = editText.getText().toString();
-                mDispatcher.dispatch(AuthenticationActionBuilder.newSendAuthEmailAction(email));
-            }
-        });
-        alert.show();
-    }
-
     // Event listeners
 
     @Subscribe(threadMode = ThreadMode.MAIN)
@@ -362,25 +323,6 @@ public class MainFragment extends Fragment {
         if (mSiteStore.hasSite()) {
             SiteModel firstSite = mSiteStore.getSites().get(0);
             prependToLog("First site name: " + firstSite.getName() + " - Total sites: " + mSiteStore.getSitesCount());
-        }
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onNewUserValidated(OnNewUserCreated event) {
-        String message = event.dryRun ? "validation" : "creation";
-        if (event.isError()) {
-            prependToLog("New user " + message + ": error: " + event.error.type + " - " + event.error.message);
-        } else {
-            prependToLog("New user " + message + ": success!");
-        }
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void onAuthEmailSent(AccountStore.OnAuthEmailSent event) {
-        if (event.isError()) {
-            prependToLog("Error sending magic link: " + event.error.type + " - " + event.error.message);
-        } else {
-            prependToLog("Sent magic link e-mail!");
         }
     }
