+++ example/src/main/res/layout/fragment_main.xml
@@ -26,10 +26,10 @@
             android:layout_alignParentEnd="true" />
 
         <Button
-            android:id="@+id/new_account"
+            android:id="@+id/signed_out_actions"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Create New Account"
+            android:text="Signed Out Actions"
             android:layout_below="@+id/sign_in_fetch_sites_button"
             android:layout_alignParentLeft="true"
             android:layout_alignParentStart="true" />
@@ -70,17 +70,5 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Media" />
-
-    <ImageView
-        android:id="@+id/divider2"
-        android:layout_width="match_parent"
-        android:layout_height="2dp"
-        android:background="@color/material_grey_600" />
-
-    <Button
-        android:id="@+id/magic_link_email"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Send Magic Link E-mail" />
 
 </LinearLayout>
